[office-ui upgrade] Move combobox to fluentui (#1469)
* update


* lintfix


* update tslint


* checkbox update


* lintfix


* update
