+++ libs/core-ui/src/lib/components/AxisConfigDialog.styles.ts
@@ -11,7 +11,7 @@ import {
   ICalloutContentStyles
 } from "office-ui-fabric-react";
 
-import { FabricStyles } from "../../FabricStyles";
+import { FabricStyles } from "../util/FabricStyles";
 
 export interface IAxisControlDialogStyles {
   wrapper: IStyle;
