+++ libs/interpret/src/lib/MLIDashboard/Controls/GlobalExplanationTab/GlobalExplanationTab.tsx
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { IComboBoxOption, IComboBox, ComboBox } from "@fluentui/react";
 import {
   Cohort,
   JointDataset,
@@ -19,9 +20,6 @@ import {
 import { localization } from "@responsible-ai/localization";
 import { Dictionary } from "lodash";
 import {
-  ComboBox,
-  IComboBox,
-  IComboBoxOption,
   IDropdownOption,
   Dropdown,
   Text,
