+++ libs/interpret/src/lib/MLIDashboard/Controls/Scatter/DataExploration.tsx
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { ComboBox, IComboBox, IComboBoxOption } from "@fluentui/react";
 import { FabricStyles } from "@responsible-ai/core-ui";
 import { localization } from "@responsible-ai/localization";
 import { AccessibleChart, IPlotlyProperty } from "@responsible-ai/mlchartlib";
 import _ from "lodash";
-import { ComboBox, IComboBox, IComboBoxOption } from "office-ui-fabric-react";
 import React from "react";
 
 import { NoDataMessage } from "../../SharedComponents/NoDataMessage";
