+++ libs/interpret/src/lib/MLIDashboard/Controls/SinglePointFeatureImportance.tsx
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { IComboBoxOption, IComboBox, ComboBox } from "@fluentui/react";
 import {
   IExplanationContext,
   ModelTypes,
@@ -10,13 +11,7 @@ import {
 } from "@responsible-ai/core-ui";
 import { localization } from "@responsible-ai/localization";
 import _ from "lodash";
-import {
-  IDropdownOption,
-  Slider,
-  ComboBox,
-  IComboBox,
-  IComboBoxOption
-} from "office-ui-fabric-react";
+import { IDropdownOption, Slider } from "office-ui-fabric-react";
 import React from "react";
 
 import { HelpMessageDict } from "../Interfaces/IStringsParam";
