+++ libs/interpret/src/lib/MLIDashboard/Controls/WhatIfTab/WhatIfPanel.tsx
@@ -1,17 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { IComboBoxOption, IComboBox, ComboBox } from "@fluentui/react";
 import {
   IExplanationModelMetadata,
   JointDataset
 } from "@responsible-ai/core-ui";
 import { localization } from "@responsible-ai/localization";
 import {
-  ComboBox,
   Dropdown,
   FocusZone,
-  IComboBox,
-  IComboBoxOption,
   IconButton,
   IDropdownOption,
   Label,
