+++ libs/interpret/src/lib/MLIDashboard/SharedComponents/FeatureEditingTile.tsx
@@ -1,16 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { IComboBoxOption, IComboBox, ComboBox } from "@fluentui/react";
 import { FabricStyles } from "@responsible-ai/core-ui";
 import { localization } from "@responsible-ai/localization";
 import { RangeTypes } from "@responsible-ai/mlchartlib";
 import { toNumber } from "lodash";
-import {
-  TextField,
-  ComboBox,
-  IComboBox,
-  IComboBoxOption
-} from "office-ui-fabric-react";
+import { TextField } from "office-ui-fabric-react";
 import React from "react";
 
 import { featureEditingTileStyles } from "./FeatureEditingTile.styles";
@@ -88,7 +84,7 @@ export class FeatureEditingTile extends React.Component<
             allowFreeform
             ariaLabel={this.props.featureName}
             autoComplete="on"
-            options={this.options}
+            options={this.options || []}
             onChange={this.onComboSelected}
             styles={FabricStyles.defaultDropdownStyle}
           />
